backend: db migrations changes: - columns types - optimized

+++ backend/src/main/java/com/smartvizz/expenses/backend/data/repositories/WalletRepository.java
@@ -4,5 +4,5 @@ import com.smartvizz.expenses.backend.data.entities.WalletEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface WalletRepository extends JpaRepository<WalletEntity, Long>, JpaSpecificationExecutor<WalletEntity> {
+public interface WalletRepository extends JpaRepository<WalletEntity, Integer>, JpaSpecificationExecutor<WalletEntity> {
 }
